disease_detection/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic-settings==2.2.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
aiokafka==0.10.0
pillow==10.2.0
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
